Fixed config values being reset to default because of a validator error.

+++ src/main/java/net/ltxprogrammer/changed/ChangedConfig.java
@@ -35,11 +35,11 @@ public class ChangedConfig {
     }
 
     public static class Server {
-        public final ForgeConfigSpec.ConfigValue<Float> transfurTolerance;
+        public final ForgeConfigSpec.ConfigValue<Double> transfurTolerance;
 
         public Server(ForgeConfigSpec.Builder builder) {
             builder.comment("Replacing gamerule transfurTolerance, this config value acts like setting the max health of the player.");
-            transfurTolerance = builder.define("transfurTolerance", 20.0f);
+            transfurTolerance = builder.define("transfurTolerance", 20.0);
         }
     }
